content/media/webaudio/test/test_convolverNodePassThrough.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test ConvolverNode with passthrough</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var gTest = {
  length: 2048,
  numberOfChannels: 1,
  createGraph: function(context) {
    var source = context.createBufferSource();

    var convolver = context.createConvolver();

    source.buffer = this.buffer;
    convolver.buffer = this.buffer;

    source.connect(convolver);

    var convolverWrapped = SpecialPowers.wrap(convolver);
    ok("passThrough" in convolverWrapped, "ConvolverNode should support the passThrough API");
    convolverWrapped.passThrough = true;

    source.start(0);
    return convolver;
  },
  createExpectedBuffers: function(context) {
    this.buffer = context.createBuffer(1, 2048, context.sampleRate);
    for (var i = 0; i < 2048; ++i) {
      this.buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
    }

    return [this.buffer];
  },
};

runTest();

</script>
</pre>
</body>
</html>
